fixed hamburger toggle / resource width
